Clamp a huge SVLEN or LEN to an int and compare a Number END as a long
longestSvlen could return 0 for Long.MIN_VALUE because Math.abs stays
negative, and both helpers could produce values whose addition to pos
overflowed a long before the clamp in furthest(). Both now cap the
magnitude at Integer.MAX_VALUE. validateStop() compares the END
attribute as a long so that a Long past the int range is rejected
instead of wrapping through intValue().

+++ src/main/java/htsjdk/variant/variantcontext/VariantContext.java
@@ -1529,9 +1529,9 @@ public class VariantContext implements HtsRecord, Feature, Serializable {
     private void validateStop() {
         if (hasAttribute(VCFConstants.END_KEY)) {
             final Object endValue = getAttribute(VCFConstants.END_KEY);
-            final int end;
+            final long end;
             if (endValue instanceof Number) {
-                end = ((Number) endValue).intValue();
+                end = ((Number) endValue).longValue();
             } else if (VCFConstants.MISSING_VALUE_v4.equals(endValue.toString())) {
                 return; // nothing to compare with
             } else {
